+++ fiddles/angular2/fiddle-0037-ModalPopUp/app/components/grid/grid.component.ts
@@ -18,8 +18,12 @@ export class GridComponent {
      constructor() {
      }
 
-     private onRowSelected($event) {
+     private onRowSelected($event):void {
           this.rowselected.emit($event.node);
+     }
+
+     private onCellValueChanged($event):void {
+          console.log($event);
      }
 
      get height():number {